Separate the steps into separate files
